Beginning changes for new OLIscag inputs.
Prior to this change we were reading legacy UCSB matlab files.
This is the beginning of changes to begin using CU-Boulder
OLI-scag files.  Wrote input parser for landsatFSCA.m. Added
start and stop file numbers to batch processor to control which
files are reprojected. Fixed location of landsat data files on PL.