Fixed some bugs on integration.
Saving a copy of STC object from interp file, because matfile (v7.3) object
throws an error when you try to reference structure fields directly.

Fixed bug that was trying to do thresholding on the albedo variables that are
not in the mosaic yet.

Fixed syntax bug on call to writeFiles method.

Limited parpool processing to 3 months (trial-and-error determined each month is
taking ~50GB memory). The bash scripts need to ask for 32 tasks.