Removed STC.minViewableSCAForFillNaN.
It is not needed, we should be using sthreshForGS and sthreshForRF.

Changed final range checks at end of Filter_SCAGDRFS to use Sgap (viewable
snow), not SgapV (adjusted snow).  Added verbosity to see if these checks are
ever actually needed, given that we are using Sgap in the calls to FillNaN3
earlier in this function.